keycloak+users

+++ src/main/resources/db/migration/V3__added_roles_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE if not exists roles
 (
-    id   BIGINT primary key auto_increment,
+    id   smallint primary key auto_increment,
     role varchar(30)
 );
